requirements.txt
pyrogram
tgcrypto
asyncio
pyromod
python_dotenv